Partial construction of Scala Plan from TS Plan
I changed the class structure and naming to more closely match what is in TypeScript. In some places I turned traits into actual case classes. I’m open to feedback on those changes.

I’m not sure how to construct the Scala TreeNode, there are so many subclasses and no clear way to determine the type (as opposed to Instructions with their InstructionKind). In TS the TreeNodes are all interfaces with abstract functions.  It is not easy to map the type or function value to the Scala world.

I probably need a way to make only the Respond Instruction a Callback, and that may mean modeling InstructionKind as sealed case object rather than an Enumeration.