Map common production names to friendlier aliases
As per #246, writing path expressions against ANTLR backed
language extensions isn't thet friendly. It requires the user
knows the grammar to navigate. Providing aliases for common
operations is a rather nice way to write clear and readable
path expressions.

Thanks to the `AstNodeCreationStrategy` trait, it's actually
rather straigtforward to achieve. But there is a catch.
Grammars aren't all equal in the granularity they expose.
For instance, the Python grammar doesn't differentiate between
methods and functions (methods are bindable functions).
Probably, we could simply expose a Func alias on the Python
side of course.

Now, let's assume we want to write a predicate for the
function's name, this cannot be easily shared because, in Python
this looks like this
`/PythonFile()//Func()[/NAME[@value='echo']]`, while in Java
it looks like this
`/JavaFile()//Func()[/methodHeader/methodDeclarator/Identifier[@value='twoArgs']]`.
I'm not sure how we could map those location paths in the
predicate so they become generic.

Finally, I wanted to map the `ScalaFileType` because I
didn't know where to declare that mapping.

Overall, this is a little more limited than expected but
there is promise if we can make the aliasing more intelligent
about the structure, not just the local node.